moving download tests
