+++ hi-ml-cpath/testhisto/testhisto/utils/test_report_utils.py
@@ -7,9 +7,7 @@ import numpy as np
 import pandas as pd
 import pandas.testing
 import pytest
-import time
 
-from health_azure.utils import download_file_if_necessary
 from health_cpath.utils.output_utils import (AML_LEGACY_TEST_OUTPUTS_CSV, AML_OUTPUTS_DIR, AML_TEST_OUTPUTS_CSV,
                                              AML_VAL_OUTPUTS_CSV)
 from health_cpath.utils.report_utils import (collect_hyperdrive_metrics, collect_hyperdrive_outputs,
@@ -72,43 +70,6 @@ def test_hyperdrive_runs_have_val_and_test_outputs() -> None:
     with patch.object(parent_run, 'get_children', return_value=[legacy_run, run_with_val_and_test]):
         with pytest.raises(ValueError, match="has mixed children"):
             child_runs_have_val_and_test_outputs(parent_run)
-
-
-@pytest.mark.parametrize('overwrite', [False, True])
-def test_download_from_run_if_necessary(tmp_path: Path, overwrite: bool) -> None:
-    filename = "test_output.csv"
-    download_dir = tmp_path
-    remote_filename = "outputs/" + filename
-    expected_local_path = download_dir / filename
-
-    def create_mock_file(name: str, output_file_path: str, _validate_checksum: bool) -> None:
-        output_path = Path(output_file_path)
-        # This odd-looking construction is to diagnose weird behaviour on the build agents,
-        # where the file doesn't seem to be created.
-        print(f"Writing mock content to file {output_path}")
-        output_path.write_text("mock content")
-        if output_path.is_file():
-            print(f"File {output_path} exists")
-        else:
-            time.sleep(2)
-            assert output_path.is_file(), "File still not on disk?"
-
-    run = MagicMock()
-    run.download_file.side_effect = create_mock_file
-
-    local_path = download_file_if_necessary(run, remote_filename, expected_local_path)
-    run.download_file.assert_called_once()
-    assert local_path == expected_local_path
-    assert local_path.exists()
-
-    run.reset_mock()
-    new_local_path = download_file_if_necessary(run, remote_filename, expected_local_path, overwrite=overwrite)
-    if overwrite:
-        run.download_file.assert_called_once()
-    else:
-        run.download_file.assert_not_called()
-    assert new_local_path == local_path
-    assert new_local_path.exists()
 
 
 class MockChildRun:
